pyproject.toml
[project]
name = "doodle-better"
version = "1.0.0"
description = "A frictionless weekly poll/availability webapp for sports pickup games"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "aiosqlite>=0.20.0",
    "pytz>=2024.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
